Issue/342/separate theory data cosmology (#344)
refactor: move theory, dataops and cosmology in separate directories

* mv eval_sigma_crit from Modeling to Cosmology

* create and use _patch_rho_crit_to_cd2018 in ct and ccl

* put py files in separated cosmology/modeling/dataops directories

* rename clmm_cosmo, clmm_modeling -> parent_class

* rename modeling -> theory

* create plotting dir

* rename unittest files

* update notebooks

* fix bug in imporing theory

* fix import im plotting

* update documentation

* add cosmology to docs

* mv support to clmm

* fix theory/__init__.py for notebooks

* update notebooks to import clmm.support

* add missing clmm/support/__init__.py file

* fix bug in docstring of clmm/support/mock_data.py

* fix bug in docstring of clmm/support/mock_data.py

* add support to docs/doc-config.ini

* rm lsst module

* Update version number
