+++ clmm/__init__.py
@@ -3,10 +3,8 @@ from .gcdata import GCData
 from .galaxycluster import GalaxyCluster
 from .dataops import compute_tangential_and_cross_components, make_radial_profile
 from .utils import compute_radial_averages, make_bins, convert_units
-from .modeling import get_reduced_shear_from_convergence, get_3d_density, predict_surface_density, predict_excess_surface_density, get_critical_surface_density, predict_tangential_shear, predict_convergence, predict_reduced_tangential_shear, Modeling, Cosmology
-from .plotting import plot_profiles
-
-from . import lsst
+from .theory import get_reduced_shear_from_convergence, get_3d_density, predict_surface_density, predict_excess_surface_density, get_critical_surface_density, predict_tangential_shear, predict_convergence, predict_reduced_tangential_shear, Modeling, Cosmology
+from . import support
 
 
-__version__ = '0.7.0'
+__version__ = '0.8.0'
