--- clmm/lsst/__init__.py
@@ -1,1 +1,0 @@
-from .load_clusters_with_gcr import *
