+++ clmm/theory/cluster_toolkit.py
@@ -12,23 +12,14 @@ from .. constants import Constants as const
 from . import func_layer
 from . func_layer import *
 
-from .. clmm_cosmo import CLMMCosmology
-from .. clmm_modeling import CLMModeling
+from .parent_class import CLMModeling
+
+from .. utils import _patch_rho_crit_to_cd2018
+
+from .. cosmology.cluster_toolkit import AstroPyCosmology
+Cosmology = AstroPyCosmology
 
 __all__ = ['CTModeling', 'Modeling', 'Cosmology']+func_layer.__all__
-
-
-def _patch_ct_to_cd2018():
-    r""" Convertion factor from cluster_toolkit hardcoded rho_crit to
-    CODATA 2018+IAU 2015
-
-    """
-
-    rhocrit_mks = 3.0*100.0*100.0/(8.0*np.pi*const.GNEWT.value)
-    rhocrit_cd2018 = rhocrit_mks*1000.0*1000.0*const.PC_TO_METER.value*1.0e6/const.SOLAR_MASS.value
-    rhocrit_cltk = 2.77533742639e+11
-
-    return rhocrit_cd2018/rhocrit_cltk
 
 
 def _assert_correct_type_ct(a):
@@ -61,7 +52,7 @@ class CTModeling(CLMModeling):
         self.mdef_dict = {'mean': 'mean'}
         self.hdpm_dict = {'nfw': 'nfw'}
         # Attributes exclusive to this class
-        self.cor_factor = _patch_ct_to_cd2018()
+        self.cor_factor = _patch_rho_crit_to_cd2018(2.77533742639e+11)
         # Set halo profile and cosmology
         self.set_halo_density_profile(halo_profile_model, massdef, delta_mdef)
         self.set_cosmo(None)
@@ -82,20 +73,6 @@ class CTModeling(CLMModeling):
 
     def set_mass(self, mdelta):
         self.mdelta = mdelta
-
-    def eval_sigma_crit(self, z_len, z_src):
-        if np.any(np.array(z_src)<=z_len):
-            warnings.warn(f'Some source redshifts are lower than the cluster redshift. Returning Sigma_crit = np.inf for those galaxies.')
-        # Constants
-        clight_pc_s = const.CLIGHT_KMS.value*1000./const.PC_TO_METER.value
-        gnewt_pc3_msun_s2 = const.GNEWT.value*const.SOLAR_MASS.value/const.PC_TO_METER.value**3
-
-        d_l = self.cosmo.eval_da_z1z2(0, z_len)
-        d_s = self.cosmo.eval_da_z1z2(0, z_src)
-        d_ls = self.cosmo.eval_da_z1z2(z_len, z_src)
-
-        beta_s = np.maximum(0., d_ls/d_s)
-        return clight_pc_s**2/(4.0*np.pi*gnewt_pc3_msun_s2)*1/d_l*np.divide(1., beta_s)*1.0e6
 
     def eval_density(self, r3d, z_cl):
         h = self.cosmo['h']
@@ -156,60 +133,4 @@ class CTModeling(CLMModeling):
         return 1./((1-kappa)**2-abs(gamma_t)**2)
 
 
-# CLMM Cosmology object - clustertoolkit+astropy implementation
-
-
-class AstroPyCosmology(CLMMCosmology):
-
-    def __init__(self, **kwargs):
-        super(AstroPyCosmology, self).__init__(**kwargs)
-
-        # this tag will be used to check if the cosmology object is accepted by the modeling
-        self.backend = 'ct'
-
-        assert isinstance(self.be_cosmo, LambdaCDM)
-
-    def _init_from_cosmo(self, be_cosmo):
-
-        assert isinstance(be_cosmo, LambdaCDM)
-        self.be_cosmo = be_cosmo
-
-    def _init_from_params(self, H0, Omega_b0, Omega_dm0, Omega_k0):
-
-        Om0 = Omega_b0+Omega_dm0
-        Ob0 = Omega_b0
-        Ode0 = 1.0-Om0-Omega_k0
-
-        self.be_cosmo = LambdaCDM(H0=H0, Om0=Om0, Ob0=Ob0, Ode0=Ode0)
-
-    def _set_param(self, key, value):
-        raise NotImplementedError("Astropy do not support changing parameters")
-
-    def _get_param(self, key):
-        if key == "Omega_m0":
-            return self.be_cosmo.Om0
-        elif key == "Omega_b0":
-            return self.be_cosmo.Ob0
-        elif key == "Omega_dm0":
-            return self.be_cosmo.Odm0
-        elif key == "Omega_k0":
-            return self.be_cosmo.Ok0
-        elif key == 'h':
-            return self.be_cosmo.H0.to_value()/100.0
-        elif key == 'H0':
-            return self.be_cosmo.H0.to_value()
-        else:
-            raise ValueError(f"Unsupported parameter {key}")
-
-    def get_Omega_m(self, z):
-        return self.be_cosmo.Om(z)
-
-    def get_E2Omega_m(self, z):
-        return self.be_cosmo.Om(z)*(self.be_cosmo.H(z)/self.be_cosmo.H0)**2
-
-    def eval_da_z1z2(self, z1, z2):
-        return self.be_cosmo.angular_diameter_distance_z1z2(z1, z2).to_value(units.Mpc)
-
-
 Modeling = CTModeling
-Cosmology = AstroPyCosmology
