+++ clmm/theory/generic.py
