+++ docs/doc-config.ini
@@ -1,19 +1,18 @@
 APIDOC
 constants
+cosmology
 dataops
 galaxycluster
 gcdata
-lsst
-modeling
+theory
 plotting
 utils
-clmm_modeling
-clmm_cosmo
+support
 
 DEMO
 ../examples/demo_generate_mock_cluster.ipynb
 ../examples/demo_dataops_functionality.ipynb
-../examples/demo_modeling_functionality.ipynb
+../examples/demo_theory_functionality.ipynb
 ../examples/DC2_examples/data_and_model_demo_DC2.ipynb
 
 EXAMPLE
