+++ examples/notebook_install.py
